app/azure_rag_api/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
azure-identity==1.7.0
azure-search-documents==11.3.0
openai==0.27.0
python-dotenv==0.20.0